ps2

+++ include/mgba/internal/arm/isa-thumb.h
@@ -12,7 +12,7 @@ CXX_GUARD_START
 
 struct ARMCore;
 
-typedef void (*ThumbInstruction)(struct ARMCore*, unsigned opcode);
+typedef void (*ThumbInstruction)(struct ARMCore*, uint16_t opcode);
 extern const ThumbInstruction _thumbTable[0x400];
 
 CXX_GUARD_END
